+++ loop_test.go
@@ -20,7 +20,7 @@ func genericForTest(t *testing.T, name string, loop ParallelForLoop) {
 	for i := uint(1); i < maxsize; i *= 2 {
 		data := makeBoolData(i)
 
-		ForInterleaved(0, len(data), func(idx int) {
+		ForInterleaved(0, uint(len(data)), func(idx uint) {
 			data[idx] = true
 		})
